requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
joblib>=1.3.0
reportlab>=4.0.0
